+++ lib/lancelot.rb
@@ -2,8 +2,8 @@
 
 require_relative "lancelot/version"
 require_relative "lancelot/lancelot"
+require_relative "lancelot/dataset"
 
 module Lancelot
   class Error < StandardError; end
-  # Your code goes here...
 end